Servicios seleccionados

+++ src/logic/Services.java
@@ -2,6 +2,6 @@ package logic;
 
 public interface Services {
     public String getNombreSer();
-    public int getPrecioSer();
+    public double getPrecioSer();
      
 }
